# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BenchBoard consultant availability and matching API
 * OpenAPI spec version: 0.1.0
 */
/**
 * @nullable
 */
export type CurrentProfileRole = typeof CurrentProfileRole[keyof typeof CurrentProfileRole] | null;


export const CurrentProfileRole = {
  consultant: 'consultant',
  client: 'client',
} as const;

export type AvailabilityStatus = typeof AvailabilityStatus[keyof typeof AvailabilityStatus];


export const AvailabilityStatus = {
  available: 'available',
  deployed: 'deployed',
  unavailable: 'unavailable',
} as const;

export interface Engagement {
  id: number;
  projectName: string;
  startDate: string;
  endDate: string;
}

export interface Consultant {
  id: number;
  name: string;
  title: string;
  skills: string[];
  serviceOffers: string[];
  hourlyRate: number;
  availabilityStatus: AvailabilityStatus;
  /** @nullable */
  engagement: Engagement | null;
  createdAt: string;
  updatedAt: string;
}

export interface Client {
  id: number;
  name: string;
  company: string;
  industry: string;
  bio: string;
  rating: number;
  createdAt: string;
  updatedAt: string;
}

export interface CurrentProfile {
  /** @nullable */
  role: CurrentProfileRole;
  profile: Consultant | Client | null;
}

export type ProfileSetupInputRole = typeof ProfileSetupInputRole[keyof typeof ProfileSetupInputRole];


export const ProfileSetupInputRole = {
  consultant: 'consultant',
  client: 'client',
} as const;

export interface ProfileSetupInput {
  role: ProfileSetupInputRole;
  /** @minLength 1 */
  name: string;
  title?: string;
  /** @minimum 0 */
  hourlyRate?: number;
  company?: string;
  industry?: string;
  bio?: string;
}

export type DemoAccountRole = typeof DemoAccountRole[keyof typeof DemoAccountRole];


export const DemoAccountRole = {
  consultant: 'consultant',
  client: 'client',
} as const;

export interface DemoAccount {
  role: DemoAccountRole;
  email: string;
  password: string;
}

export interface DemoAccounts {
  accounts: DemoAccount[];
}

export interface HireInput {
  /** @minimum 1 */
  consultantId: number;
}

export type HireResponseEngagement = { [key: string]: unknown };

export interface HireResponse {
  message: string;
  engagement: HireResponseEngagement;
}

export interface HealthStatus {
  status: string;
}

export interface ConsultantInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  title: string;
  /** @minItems 1 */
  skills: string[];
  serviceOffers?: string[];
  /** @minimum 0 */
  hourlyRate: number;
  availabilityStatus: AvailabilityStatus;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  engagementStartDate?: string | null;
  /** @nullable */
  engagementEndDate?: string | null;
}

export interface ConsultantUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  title?: string;
  /** @minItems 1 */
  skills?: string[];
  serviceOffers?: string[];
  /** @minimum 0 */
  hourlyRate?: number;
  availabilityStatus?: AvailabilityStatus;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  engagementStartDate?: string | null;
  /** @nullable */
  engagementEndDate?: string | null;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  industry: string;
  bio?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  company?: string;
  /** @minLength 1 */
  industry?: string;
  bio?: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  planning: 'planning',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface Project {
  id: number;
  clientId: number;
  name: string;
  description: string;
  skills: string[];
  status: ProjectStatus;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  clientId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 20 */
  description: string;
  skills: string[];
  status: ProjectStatus;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 20 */
  description?: string;
  skills?: string[];
  status?: ProjectStatus;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface DashboardSummary {
  totalConsultants: number;
  available: number;
  deployed: number;
  unavailable: number;
  endingSoon: number;
}

export interface MatchInput {
  /** @minLength 20 */
  jobDescription: string;
}

export interface MatchRequirements {
  skills: string[];
  /** @nullable */
  seniority: string | null;
  /** @nullable */
  domain: string | null;
}

export interface ConsultantMatch {
  rank: number;
  consultant: Consultant;
  matchingSkills: string[];
  reason: string;
  score: number;
}

export interface MatchResponse {
  requirements: MatchRequirements;
  matches: ConsultantMatch[];
}

export interface ErrorResponse {
  error: string;
}

export type ListConsultantsParams = {
search?: string;
status?: AvailabilityStatus;
};

export type ListProjectsParams = {
clientId: number;
};