# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BenchBoard consultant availability and matching API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultants:
    get:
      operationId: listConsultants
      tags: [consultants]
      summary: List consultants
      description: Returns the consultant roster with active engagement details.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AvailabilityStatus"
      responses:
        "200":
          description: Consultant roster
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultant"
    post:
      operationId: createConsultant
      tags: [consultants]
      summary: Add a consultant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultantInput"
      responses:
        "201":
          description: Created consultant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
        "400":
          description: Invalid consultant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consultants/{id}:
    patch:
      operationId: updateConsultant
      tags: [consultants]
      summary: Update a consultant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultantUpdate"
      responses:
        "200":
          description: Updated consultant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
        "404":
          description: Consultant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteConsultant
      tags: [consultants]
      summary: Remove a consultant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Consultant removed
        "404":
          description: Consultant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      description: Returns client profiles ordered by rating.
      responses:
        "200":
          description: Client profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AvailabilityStatus:
      type: string
      enum: [available, deployed, unavailable]
    Engagement:
      type: object
      properties:
        id:
          type: integer
        projectName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [id, projectName, startDate, endDate]
    Consultant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        skills:
          type: array
          items:
            type: string
        serviceOffers:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
        availabilityStatus:
          $ref: "#/components/schemas/AvailabilityStatus"
        engagement:
          type: ["object", "null"]
          $ref: "#/components/schemas/Engagement"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, title, skills, serviceOffers, hourlyRate, availabilityStatus, engagement, createdAt, updatedAt]
    ConsultantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        skills:
          type: array
          minItems: 1
          items:
            type: string
        serviceOffers:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          minimum: 0
        availabilityStatus:
          $ref: "#/components/schemas/AvailabilityStatus"
        projectName:
          type: ["string", "null"]
        engagementStartDate:
          type: ["string", "null"]
          format: date
        engagementEndDate:
          type: ["string", "null"]
          format: date
      required: [name, title, skills, hourlyRate, availabilityStatus]
    ConsultantUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        skills:
          type: array
          minItems: 1
          items:
            type: string
        serviceOffers:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          minimum: 0
        availabilityStatus:
          $ref: "#/components/schemas/AvailabilityStatus"
        projectName:
          type: ["string", "null"]
        engagementStartDate:
          type: ["string", "null"]
          format: date
        engagementEndDate:
          type: ["string", "null"]
          format: date
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        industry:
          type: string
        bio:
          type: string
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, company, industry, bio, rating, createdAt, updatedAt]
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        company:
          type: string
          minLength: 1
        industry:
          type: string
          minLength: 1
        bio:
          type: string
      required: [name, company, industry]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        company:
          type: string
          minLength: 1
        industry:
          type: string
          minLength: 1
        bio:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]